requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
boto3==1.34.11
botocore==1.34.162
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap5==0.7
cryptography==44.0.2
decorator==5.2.1
defusedxml==0.7.1
Django==4.2.11
django-allauth==0.57.0
django-crispy-forms==2.0
django-environ==0.10.0
django-filter==23.3
django-phonenumber-field==7.1.0
django-qr-code==3.1.1
django-rest-framework==0.1.0
django-simple-history==3.3.0
django-storages==1.14.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
exceptiongroup==1.2.2
executing==2.2.0
gunicorn==21.2.0
idna==3.10
importlib_metadata==8.6.1
ipython==8.18.1
jedi==0.19.2
jmespath==1.0.1
matplotlib-inline==0.1.7
oauthlib==3.2.2
openai==1.26.0
packaging==24.2
parso==0.8.4
paypalrestsdk==1.13.1
pexpect==4.9.0
phonenumbers==8.13.11
Pillow==10.0.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
pypng==0.20220715.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2025.2
qrcode==7.4.2
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
segno==1.6.6
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
stripe==5.4.0
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==1.26.20
wcwidth==0.2.13
zipp==3.21.0
